Add facilities for fake loading progress

+++ CodingChallenge/Common/Networking/URLSession+Load.swift
@@ -69,7 +69,7 @@ extension URLSession.DataTaskLoadPublisher {
 
             progressObservation = dataTask.progress
                 .observe(\.fractionCompleted) { progress, _ in
-                    _ = downstream.receive(.pending(Float(progress.fractionCompleted)))
+                    _ = downstream.receive(.pending(Double(progress.fractionCompleted)))
                 }
 
             dataTask.resume()
